requirements.txt
streamlit
flask
pandas
pymongo
python-dotenv
requests
plotly
numpy
streamlit_lottie
